requirements.txt
selenium
requests
opencv-python
google-cloud-storage
